perf(webp): i16 coefficient port, stage A - the quantized-levels layer (#38 speed)
Output bytes unchanged (output_hash grid identical; full suites on
x86-64 AND i686; wasm32 no_std build clean; clippy x3 + fmt).

Quantized LEVELS (|level| <= 2047) now travel as i16 end-to-end,
matching libwebp's int16_t:

- Residual::coeffs, get_residual_cost_* (the SSE2 kernel loads two
  128-bit i16 vectors directly - the i32->i16 pack prelude is GONE;
  wasm lanes widen at load), record_coeff_tokens / record_coeffs.
- trellis_quantize_block's zigzag OUT param.
- Every carrier: QuantizedMbCoeffs, LumaBlockResult::{y1_zigzag,
  y2_zigzag}, ChromaBlockResult::zigzag, I4WinnerCarry::levels_zz,
  UvWinnerCarry::zigzag_*, MacroblockInfo::intra16_y2_zz,
  I4BlockResult::levels_zz, the RD loops' zigzag views
  (y1_quant_zz/y2_quant_zz/uv_quant_zz) - halving their memory.
- is_flat_coeffs / all_levels staging conversions deleted (types now
  agree); store_max_delta takes i16.

Stage B next: quantize/dequantize block arrays and the trellis DP
internals to i16 (the m6 lever - libwebp's TrellisQuantizeBlock runs
int16_t end-to-end at ~half zen's per-call cost).

+++ src/encoder/cost/stats.rs
@@ -147,7 +147,7 @@ pub enum TokenType {
 /// * `ctx` - Initial context (0, 1, or 2)
 /// * `stats` - Statistics accumulator
 pub fn record_coeffs(
-    coeffs: &[i32],
+    coeffs: &[i16],
     token_type: TokenType,
     first: usize,
     ctx: usize,
@@ -179,7 +179,7 @@ pub fn record_coeffs(
     // Process coefficients up to end_of_block (last non-zero + 1)
     while n < end_of_block {
         let band = VP8_ENC_BANDS[n] as usize;
-        let v = coeffs[n].unsigned_abs();
+        let v = u32::from(coeffs[n].unsigned_abs());
         n += 1;
 
         // Record at node 0 (EOB check) if not skipping
